requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=15.0.0
pybind11
pylatexenc
ray[data,train,tune,serve]
tensordict<0.6
torchdata
transformers
vllm<=0.6.3
wandb
flask
tqdm
openai
argparse
sqlite3
multiprocessing
re